+++ examples/distributedactors.ts
@@ -1,8 +1,11 @@
+import { createServer } from "http";
+
 import { Actor } from "../src/Actor";
 import type { ActorRef } from "../src/ActorRef";
 import type { ActorSystem } from "../src/ActorSystem";
 import { DistributedActorSystem } from "../src/DistributedActorSystem";
-import { NatsDistributor } from "../src/NatsDistributor";
+// import { NatsDistributor } from "../src/NatsDistributor";
+import { WebsocketDistributor } from "../src/WebsocketDistributor";
 
 class Ping extends Actor<string, void|string> {
 	private counter = 0;
@@ -65,17 +68,19 @@ class Pong extends Actor<string, void> {
 }
 
 async function run() {
-    const distributor1 = new NatsDistributor("SA", "9999");
-    const distributor2 = new NatsDistributor("SB", "9999");
+    const srv = createServer().listen(12345,"127.0.0.1", async () => {
+        const distributor1 = new WebsocketDistributor("SA", { server: srv });
+        const distributor2 = new WebsocketDistributor("SB", "ws://127.0.0.1:12345");
 
-    const systemA = new DistributedActorSystem({distributor: distributor1, systemName: "SA"});
-    const ping = await systemA.createActor(Ping, { name: "PING" });
+        const systemA = new DistributedActorSystem({distributor: distributor1, systemName: "SA"}, () => { return Promise.resolve();});
+        const ping =  await systemA.createActor(Ping, { name: "PING" });
 
-    const systemB = new DistributedActorSystem({distributor: distributor2, systemName: "SB"});
-    const pong = await systemB.createActor(Pong, { name: "PONG" });
+        const systemB = new DistributedActorSystem({distributor: distributor2, systemName: "SB"}, () => { return Promise.resolve(); });
+        const pong = await systemB.createActor(Pong, { name: "PONG" });
 
-    systemA.started().then(() => systemA.send(ping, "PING"));
-    systemB.started().then(() => systemB.send(pong, "ASK"));
+        systemA.started().then(() => systemA.send(ping, "PING"));
+        systemB.started().then(() => systemB.send(pong, "ASK"));
+    });
 }
 
 run();